+++ projects/online/online/dataloading/__init__.py
@@ -1,3 +1,3 @@
-from .ngdd import data_iterator as ngdd_data_iterator
+from .arrakis import data_iterator as arrakis_data_iterator
 from .offline import offline_data_iterator
 from .online import data_iterator
